+++ comics/comics/dilbert.py
@@ -6,17 +6,17 @@ class ComicData(ComicDataBase):
     name = 'Dilbert'
     language = 'en'
     url = 'http://www.dilbert.com/'
-    start_date = '1989-04-06'
+    start_date = '1989-04-16'
     rights = 'Scott Adams'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 32
+    history_capable_date = '1989-04-16'
     schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
     time_zone = 'US/Mountain'
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://feeds.feedburner.com/DilbertDailyStrip')
-        for entry in feed.for_date(pub_date):
-            url = entry.summary.src('img[src*="dilbert.com"]')
-            return CrawlerImage(url)
+        page = self.parse_page(
+            pub_date.strftime('http://dilbert.com/strips/comic/%Y-%m-%d/'))
+        url = page.src('img[src$=".strip.zoom.gif"]')
+        return CrawlerImage(url)