+++ docs/changes.rst
@@ -39,8 +39,9 @@ v2.2.0 (in development)
 - Update: ``beyondthetree`` is no longer published.
 - Update: ``chainsawsuit`` after feed change.
 - Update: ``choppingblock`` is no longer published.
-- Update: ``dilbertvg`` after move to new site.
 - Update: ``darylcagle`` after feed change.
+- Update: ``dilbert`` after removal of feed.
+- Update: ``dilbertvg`` after move to new site.
 - Update: ``eatthattoast`` to not throw exception when the site is broken.
 - Update: ``evilinc`` to be more robust.
 - Update: ``extralife`` after feed change.